move const enums out of public-types.ts
