+++ packages-exp/auth-exp/src/core/providers/email.ts
@@ -15,9 +15,8 @@
  * limitations under the License.
  */
 
+import { ProviderId, SignInMethod } from '../../model/enums';
 import {
-  ProviderId,
-  SignInMethod,
   AuthProvider
 } from '../../model/public_types';
