+++ packages-exp/auth-exp/src/core/providers/google.ts
@@ -16,8 +16,6 @@
  */
 
 import {
-  ProviderId,
-  SignInMethod,
   UserCredential
 } from '../../model/public_types';
 import { FirebaseError } from '@firebase/util';
@@ -27,6 +25,7 @@ import { TaggedWithTokenResponse } from '../../model/id_token';
 import { UserCredentialInternal } from '../../model/user';
 import { OAuthCredential } from '../credentials/oauth';
 import { BaseOAuthProvider } from './oauth';
+import { ProviderId, SignInMethod } from '../../model/enums';
 
 /**
  * Provider for generating an an {@link OAuthCredential} for {@link ProviderId}.GOOGLE.
