+++ packages-exp/auth-exp/src/core/providers/twitter.ts
@@ -33,8 +33,6 @@
  */
 
 import {
-  ProviderId,
-  SignInMethod,
   UserCredential
 } from '../../model/public_types';
 import { FirebaseError } from '@firebase/util';
@@ -44,6 +42,7 @@ import { TaggedWithTokenResponse } from '../../model/id_token';
 import { UserCredentialInternal } from '../../model/user';
 import { OAuthCredential } from '../credentials/oauth';
 import { BaseOAuthProvider } from './oauth';
+import { ProviderId, SignInMethod } from '../../model/enums';
 
 /**
  * Provider for generating an {@link OAuthCredential} for {@link ProviderId}.TWITTER.
