+++ packages-exp/auth-exp/src/core/strategies/anonymous.ts
@@ -15,11 +15,12 @@
  * limitations under the License.
  */
 
-import { Auth, OperationType, UserCredential } from '../../model/public_types';
+import { Auth, UserCredential } from '../../model/public_types';
 import { signUp } from '../../api/authentication/sign_up';
 import { UserInternal } from '../../model/user';
 import { UserCredentialImpl } from '../user/user_credential_impl';
 import { _castAuth } from '../auth/auth_impl';
+import { OperationType } from '../../model/enums';
 
 /**
  * Asynchronously signs in as an anonymous user.
